fix: clarify deserialization policy failures
## Summary
Make the deserialization error message accurate for both missing policies and
rejected classes, and move the persisted-UDF public-path regression into a
standalone suite that replays cleanly onto supported Spark 4 branches.

## Prompting Intent
Address the suppressed exact-head review finding, triage the failed release
compatibility check from its logs, preserve the same public UDF migration proof,
and keep the MSRC fix small rather than changing branch-specific production
behavior.

## Linked Sources
- Pull request: https://github.com/microsoft/SynapseML/pull/2678
- Failed compatibility run: https://msdata.visualstudio.com/b9b2accc-2d1c-45b3-9d24-0eb5d78cc47f/_build/results?buildId=233926848

## Rationale
The prior exception text claimed every failure lacked a constraint even when a
class had been rejected by an active policy. A neutral base message lets each
call site supply the accurate reason. The Spark 4.1 job failed before compile
because its branch-specific UDF suite intentionally hoists lambdas for Scala
2.13; placing the new persistence regression in its own suite removes that
test-only replay conflict without changing coverage or production scope.
